requirements.txt
tensorflow==1.13.1
requests
pillow
imagehash
pandas